requirements.txt
langchain
langchain-openai
fastapi
uvicorn
python-dotenv
-qU "langchain-chroma>=0.1.2"
